refactor(team): own agent-roster + personas + agent-assets health checks
Migrates 3 builtin doctor checks out of src/core/doctor.ts and into
plugins/team/lib/health-checks.ts, registered in team's activate() via
ctx.registerHealthCheck. Doctor's runDiagnostics() picks them up via
runPluginHealthChecks() instead of direct calls.

Each check drops its `autoFix` parameter and reads
getSettings().doctor.autoFixSkill inline (matches workflows precedent).
Check ids gain plugin namespacing: agent-roster → team.agent-roster,
personas → team.personas, agent-assets → team.agent-assets.

Tests: tests/plugins/team/health-checks.test.ts absorbs
tests/core/doctor-agent-assets.test.ts content plus adds direct
coverage for checkAgentRoster, checkPersonas, checkAgentAssets.

doctor.ts: 1762 → 1629 lines.

Refs .claude/specs/doctor-decoupling.md, .claude/specs/doctor-decoupling-plan.md (C1).
