--- tests/core/doctor-agent-assets.test.ts
@@ -1,170 +1,0 @@
-/**
- * Doctor's agent-assets check integration (Phase I-1).
- *
- * Verifies that `runDiagnostics` exercises the agent-assets onboarding
- * component and translates its CheckResult / InstallResult into
- * DiagnosticResult shape. Drift detection + autoFix paths covered.
- *
- * The full agent-assets scan / install behavior lives in
- * tests/cli/install-agent-assets.test.ts — this test only proves the
- * doctor wiring forwards the right signals, not the underlying scan
- * mechanics.
- */
-import { tmpdir } from 'os'
-import { join as pathJoin } from 'path'
-import { randomUUID } from 'crypto'
-
-const testDir = pathJoin(tmpdir(), `bakin-test-doctor-agent-assets-${Date.now()}-${randomUUID()}`)
-const openClawDir = pathJoin(testDir, 'openclaw')
-process.env.OPENCLAW_HOME = openClawDir
-process.env.BAKIN_HOME = testDir
-
-import { describe, it, expect, beforeEach, afterAll, mock } from 'bun:test'
-import { existsSync, mkdirSync, rmSync, writeFileSync } from 'fs'
-import { join } from 'path'
-
-mock.module('@/core/content-dir', () => ({
-  getContentDir: () => testDir,
-  getBakinPaths: () => ({}),
-  isUsingBakinHome: () => true,
-}))
-mock.module('@bakin/core/content-dir', () => ({
-  getContentDir: () => testDir,
-  getBakinPaths: () => ({}),
-  isUsingBakinHome: () => true,
-}))
-mock.module('@bakin/core/openclaw-home', () => ({
-  getOpenClawHome: () => openClawDir,
-  getOpenClawPath: (...parts: string[]) => join(openClawDir, ...parts),
-  resetOpenClawHome: () => {},
-}))
-
-let openClawAgents: Array<{ id: string; identity?: { name?: string } }> = []
-mock.module('@bakin/core/openclaw-config', () => ({
-  readOpenClawConfig: () => ({ agents: { list: openClawAgents } }),
-  resetOpenClawConfigCache: () => {},
-  getAgentList: () => openClawAgents,
-  getAgentIds: () => openClawAgents.map((a) => a.id),
-  findAgentById: (id: string) => openClawAgents.find((a) => a.id === id) ?? null,
-}))
-
-const adapterMockFactory = () => ({
-  addAgent: async (input: { id: string }) => {
-    openClawAgents.push({ id: input.id, identity: { name: input.id } })
-    return { id: input.id, workspace: join(openClawDir, 'workspaces', input.id) }
-  },
-  addToAllowLists: () => {},
-  removeAgent: async () => true,
-  removeFromAllowLists: () => {},
-  getOpenClawConfig: () => ({ agents: { list: openClawAgents } }),
-  listAgents: () => [],
-  getAgentIds: () => openClawAgents.map((a) => a.id),
-})
-mock.module('@bakin/team/lib/openclaw-adapter', adapterMockFactory)
-mock.module('../../plugins/team/lib/openclaw-adapter', adapterMockFactory)
-
-import { agentAssetsComponent } from '../../src/core/onboarding/agent-assets'
-import { installPackage } from '../../src/core/agent-packages/installer'
-
-afterAll(() => {
-  rmSync(testDir, { recursive: true, force: true })
-})
-
-beforeEach(() => {
-  rmSync(testDir, { recursive: true, force: true })
-  mkdirSync(testDir, { recursive: true })
-  mkdirSync(openClawDir, { recursive: true })
-  openClawAgents = []
-})
-
-function seedAgentPackage(): string {
-  const dir = join(testDir, 'pixel-pkg')
-  mkdirSync(join(dir, 'workspace'), { recursive: true })
-  mkdirSync(join(dir, 'knowledge'), { recursive: true })
-  mkdirSync(join(dir, 'assets'), { recursive: true })
-  writeFileSync(
-    join(dir, 'bakin-package.json'),
-    JSON.stringify({
-      id: 'pixel',
-      kind: 'agent',
-      name: 'Pixel',
-      version: '0.1.0',
-      agent: { identity: { name: 'Pixel' } },
-      install: { writeWorkspaceFiles: true, enableKnowledge: ['style'] },
-      contributions: {
-        workspaceFiles: ['workspace/SOUL.md'],
-        knowledge: ['knowledge/style.md'],
-        assets: ['assets/avatar.jpg'],
-      },
-    }),
-  )
-  writeFileSync(
-    join(dir, 'workspace', 'SOUL.md'),
-    `# Soul Pixel\n\n<!-- bakin:knowledge-catalog:start -->\n<!-- bakin:knowledge-catalog:end -->\n`,
-  )
-  writeFileSync(
-    join(dir, 'knowledge', 'style.md'),
-    `---\ntitle: Style\ndefaultEnabled: true\n---\n\nStyle body.`,
-  )
-  writeFileSync(join(dir, 'assets', 'avatar.jpg'), 'jpg-bytes')
-  return dir
-}
-
-const NON_INTERACTIVE = {
-  interactive: false,
-  autoApprove: true,
-  json: false,
-  checkOnly: false,
-  force: false,
-}
-
-describe('doctor agent-assets integration (Phase I-1)', () => {
-  it('component check returns ok when projections match the lockfile', async () => {
-    const src = seedAgentPackage()
-    await installPackage({ source: src })
-
-    const result = await agentAssetsComponent.check()
-    expect(result.status).toBe('ok')
-  })
-
-  it('component check returns warn when an asset is missing or drifted', async () => {
-    const src = seedAgentPackage()
-    await installPackage({ source: src })
-
-    // Mutate avatar.jpg — non-template projection sha drift
-    const avatar = join(testDir, 'agents', 'pixel', 'avatar.jpg')
-    writeFileSync(avatar, 'corrupted-content')
-
-    const result = await agentAssetsComponent.check()
-    expect(result.status).toBe('warn')
-    expect(result.message).toContain('drifted')
-  })
-
-  it('component install in autoFix mode repairs non-template drift in-place', async () => {
-    const src = seedAgentPackage()
-    await installPackage({ source: src })
-
-    // Delete an asset projection (avatar.jpg) — non-template, so update
-    // mode repairs it. Workspace files (SOUL.md / IDENTITY.md / etc.)
-    // are templateOnly and only repaired via --refresh-template; the
-    // doctor's autoFix path doesn't pass that flag because the agent
-    // may have legitimately edited those files post-install.
-    const avatar = join(testDir, 'agents', 'pixel', 'avatar.jpg')
-    rmSync(avatar)
-    expect(existsSync(avatar)).toBe(false)
-
-    const beforeCheck = await agentAssetsComponent.check()
-    expect(beforeCheck.status).toBe('warn')
-    expect(beforeCheck.message).toContain('missing')
-
-    const installResult = await agentAssetsComponent.install(NON_INTERACTIVE)
-    expect(['installed', 'noop']).toContain(installResult.status)
-
-    // Avatar is back
-    expect(existsSync(avatar)).toBe(true)
-
-    // And the post-repair check is clean
-    const afterCheck = await agentAssetsComponent.check()
-    expect(afterCheck.status).toBe('ok')
-  })
-})
